OK-513: Refaktoroitu latausanimaatio yhteiskäyttöiseen Spinner-komponenttiin, jolle asetetaan käännettu aria-label
