+++ src/app/(root)/@controls/page.tsx
@@ -9,7 +9,6 @@ import {
   OutlinedInput,
   Box,
   InputAdornment,
-  CircularProgress,
 } from '@mui/material';
 
 import { Tila, getHakuAlkamisKaudet } from '@/app/lib/kouta-types';
@@ -18,6 +17,7 @@ import { useHakuSearchParams } from '@/app/hooks/useHakuSearch';
 import { useHakutavat } from '@/app/hooks/useHakutavat';
 import { useTranslations } from '@/app/hooks/useTranslations';
 import { OphSelectControl, OphSelect } from '@/app/components/oph-select';
+import { Spinner } from '@/app/components/spinner';
 
 const HakutapaSelect = ({
   value: selectedHakutapa,
@@ -51,9 +51,7 @@ const SelectFallback = () => (
     disabled={true}
     startAdornment={
       <InputAdornment position="start">
-        <CircularProgress
-          sx={{ height: '24px !important', width: '24px !important' }}
-        />
+        <Spinner sx={{ height: '24px !important', width: '24px !important' }} />
       </InputAdornment>
     }
   />
