+++ src/app/haku/[oid]/hakukohde-panel.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import { styled, IconButton, FormLabel, CircularProgress } from '@mui/material';
+import { styled, IconButton, FormLabel } from '@mui/material';
 import CloseIcon from '@mui/icons-material/Close';
 import KeyboardDoubleArrowRightIcon from '@mui/icons-material/KeyboardDoubleArrowRight';
 import HakukohdeList from './hakukohde-list';
@@ -9,6 +9,7 @@ import { useState } from 'react';
 import { colors } from '@/app/theme';
 import { useTranslations } from '@/app/hooks/useTranslations';
 import { QuerySuspenseBoundary } from '@/app/components/query-suspense-boundary';
+import { Spinner } from '@/app/components/spinner';
 
 const StyledPanel = styled('div')({
   width: '16vw',
@@ -45,11 +46,7 @@ export const HakukohdePanel = ({ oid }: { oid: string }) => {
   return (
     <StyledPanel className={minimized ? 'minimized' : ''}>
       {!minimized && (
-        <QuerySuspenseBoundary
-          suspenseFallback={
-            <CircularProgress aria-label={t('yleinen.ladataan')} />
-          }
-        >
+        <QuerySuspenseBoundary suspenseFallback={<Spinner />}>
           <IconButton
             sx={{ alignSelf: 'right', width: '1rem', height: '1rem' }}
             onClick={() => setMinimized(true)}
