+++ src/app/haku/[oid]/hakukohde/[hakukohde]/perustiedot/page.tsx
@@ -6,9 +6,9 @@ import BasicInfo from './basic-info';
 import { useSuspenseQueries } from '@tanstack/react-query';
 import { ValintatapajonotTable } from './valintatapajonot-table';
 import { getHaku } from '@/app/lib/kouta';
-import { CircularProgress } from '@mui/material';
 import { useTranslations } from '@/app/hooks/useTranslations';
 import { QuerySuspenseBoundary } from '@/app/components/query-suspense-boundary';
+import { Spinner } from '@/app/components/spinner';
 
 type PerustiedotParams = {
   oid: string;
@@ -60,11 +60,7 @@ export default function PerustiedotTab({
     <TabContainer>
       <BasicInfo hakukohdeOid={params.hakukohde} />
       <h3>{t('perustiedot.taulukko.otsikko')}</h3>
-      <QuerySuspenseBoundary
-        suspenseFallback={
-          <CircularProgress aria-label={t('yleinen.ladataan')} />
-        }
-      >
+      <QuerySuspenseBoundary suspenseFallback={<Spinner />}>
         <PerustiedotContent oid={params.oid} hakukohde={params.hakukohde} />
       </QuerySuspenseBoundary>
     </TabContainer>
