+++ src/app/haku/[oid]/loading.tsx
@@ -1,4 +1,4 @@
-import { FullSpinner } from '@/app/components/full-spinner';
+import { FullSpinner } from '@/app/components/spinner';
 
 export default function Loading() {
   return <FullSpinner />;
