+++ src/app/localization-provider.tsx
@@ -1,6 +1,6 @@
 'use client';
 import { I18nextProvider } from 'react-i18next';
-import { FullSpinner } from './components/full-spinner';
+import { FullSpinner } from './components/spinner';
 import { useAsiointiKieli } from './hooks/useAsiointiKieli';
 import { createLocalization } from './lib/localization/localizations';
 import { ErrorView } from './components/error-view';
