+++ src/app/permission-provider.tsx
@@ -1,6 +1,6 @@
 'use client';
 import { ErrorView } from './components/error-view';
-import { FullSpinner } from './components/full-spinner';
+import { FullSpinner } from './components/spinner';
 import { useQueryUserPermissions } from './hooks/useUserPermissions';
 
 export default function PermissionProvider({
